refactor(cli): split hooks report rendering out of hooks-cli
The new toggle diagnostics pushed src/cli/hooks-cli.ts to 709 lines, over
the 700-line cap. Move the hooks list/info/check renderers and their
formatting helpers to hooks-cli.format.ts so command wiring and report
rendering each stay readable; no suppression added.
